requirements.txt
lpips==0.1.4
opencv-python==4.10.0.84
scikit-image==0.24.0
pandas==2.2.2
PyYAML==6.0.2
tqdm==4.66.4
matplotlib==3.9.0
pillow==10.2.0
numpy==1.26.4
gfpgan==1.3.8
facexlib==0.3.0
realesrgan==0.3.0
torch==1.13.1+cpu
torchvision==0.14.1+cpu
